2nd

+++ dataset/HCP_3sets_normXY.py
@@ -62,9 +62,9 @@ class Retinotopy(InMemoryDataset):
 
             data_list.append(data)
 
-        train=data_list[0:round(len(data_list)*0.6)]
-        dev=data_list[round(len(data_list)*0.6):round(len(data_list)*0.8)]
-        test=data_list[round(len(data_list)*0.8):len(data_list)]
+        train = data_list[0:int(round(len(data_list) * 0.6))]
+        dev = data_list[int(round(len(data_list) * 0.6)):int(round(len(data_list) * 0.8))]
+        test = data_list[int(round(len(data_list) * 0.8)):len(data_list)]
 
         torch.save(self.collate(train),self.processed_paths[0])
         torch.save(self.collate(dev), self.processed_paths[1])
